// src/modules/promotion/promotion.enum.ts
export enum PromotionType {
  PERCENT = "PERCENT",
  FIXED = "FIXED",
}

export enum PromotionFieldName {
  FRANCHISE_ID = "franchise_id",
  FRANCHISE_NAME = "franchise_name",
  PRODUCT_FRANCHISE_ID = "product_franchise_id",
  TYPE = "type",
  VALUE = "value",
  START_DATE = "start_date",
  END_DATE = "end_date",
  CREATED_BY = "created_by",
  IS_ACTIVE = "is_active",
  IS_DELETED = "is_deleted",
}